+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Pressing enter in app drawer launches the first app in search results ([#331])
 
+### Fixed
+- Prevent duplicated widget rendering on the home screen ([#320])
+
 ## [1.9.0] - 2026-02-02
 ### Added
 - Option to hide app labels ([#55])